Add parameter doc and moved BackstageAppButton component to its own file
