+++ common/api/appui-react.api.md
@@ -602,10 +602,8 @@ export function BackstageAppButton(props: BackstageAppButtonProps): JSX.Element;
 
 // @public
 export interface BackstageAppButtonProps {
-    // (undocumented)
     execute?: () => void;
     icon?: string;
-    // (undocumented)
     label?: string;
 }
