+++ ui/appui-react/src/appui-react/widgets/BasicToolWidget.tsx
@@ -14,7 +14,8 @@ import { SelectionContextToolDefinitions } from "../selection/SelectionContextIt
 import { ToolbarComposer } from "../toolbar/ToolbarComposer";
 import { ToolbarHelper } from "../toolbar/ToolbarHelper";
 import { UiFramework, UiVisibilityEventArgs } from "../UiFramework";
-import { BackstageAppButton, ToolWidgetComposer } from "./ToolWidgetComposer";
+import { ToolWidgetComposer } from "./ToolWidgetComposer";
+import { BackstageAppButton } from "./BackstageAppButton";
 
 /** Properties that can be used to append items to the default set of toolbar items of [[ReviewToolWidget]].
  * @public
