+++ ui/appui-react/src/test/widgets/ToolWidgetComposer.test.tsx
@@ -9,7 +9,8 @@ import * as React from "react";
 import { Provider } from "react-redux";
 import { UiFramework } from "../../appui-react";
 import { FrameworkVersion } from "../../appui-react/hooks/useFrameworkVersion";
-import { BackstageAppButton, ToolWidgetComposer } from "../../appui-react/widgets/ToolWidgetComposer";
+import { ToolWidgetComposer } from "../../appui-react/widgets/ToolWidgetComposer";
+import { BackstageAppButton } from "../../appui-react/widgets/BackstageAppButton";
 import TestUtils, { mount, storageMock } from "../TestUtils";
 import { IModelApp, NoRenderApp } from "@itwin/core-frontend";
 import { expect } from "chai";
